LPS-75598 Split UAD work into the appropriate modules - api and service
