+++ modules/apps/foundation/user-associated-data/user-associated-data-api/bnd.bnd
@@ -1,0 +1,6 @@
+Bundle-Name: Liferay User Associated Data API
+Bundle-SymbolicName: com.liferay.user.associated.data.api
+Bundle-Version: 1.0.0
+Export-Package:\
+	com.liferay.user.associated.data.model,\
+	com.liferay.user.associated.data.model.impl
