+++ modules/apps/foundation/user-associated-data/user-associated-data-api/src/main/resources/com/liferay/user/associated/data/model/impl/packageinfo
