+++ modules/apps/foundation/user-associated-data/user-associated-data-service/src/main/java/com/liferay/user/associated/data/service/registry/UADServiceRegistry.java
@@ -1,0 +1,94 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.user.associated.data.service.registry;
+
+import com.liferay.osgi.service.tracker.collections.map.ServiceReferenceMapper;
+import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMap;
+import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMapFactory;
+import com.liferay.portal.kernel.util.Validator;
+import com.liferay.user.associated.data.model.UADAsset;
+import com.liferay.user.associated.data.model.UADService;
+
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+
+import org.osgi.framework.BundleContext;
+import org.osgi.framework.ServiceReference;
+import org.osgi.service.component.annotations.Activate;
+import org.osgi.service.component.annotations.Component;
+
+/**
+* @author William Newbury
+*/
+@Component(immediate = true, service = UADServiceRegistry.class)
+public class UADServiceRegistry {
+
+	public UADService getUADService(String key) {
+		return _uadServiceTrackerMap.getService(key);
+	}
+
+	public Set<String> getUADServiceKeySet() {
+		return _uadServiceTrackerMap.keySet();
+	}
+
+	public Collection<UADService> getUADServices() {
+		return _uadServiceTrackerMap.values();
+	}
+
+	public void notify(long userId) {
+		System.out.println("############################ notify");
+
+		Map<String, List<UADAsset>> uadAssetsMap = new HashMap<>();
+
+		for (String key : getUADServiceKeySet()) {
+			UADService uadService = getUADService(key);
+
+			uadAssetsMap.put(key, uadService.getUADAssets(userId));
+		}
+	}
+
+	@Activate
+	protected void activate(BundleContext bundleContext) {
+		_uadServiceTrackerMap = ServiceTrackerMapFactory.openSingleValueMap(
+			bundleContext, UADService.class, "(uad.service.key=*)",
+			new UADServiceReferenceMapper());
+	}
+
+	private ServiceTrackerMap<String, UADService> _uadServiceTrackerMap;
+
+	private class UADServiceReferenceMapper
+		implements ServiceReferenceMapper<String, UADService> {
+
+		@Override
+		public void map(
+			ServiceReference<UADService> serviceReference,
+			Emitter<String> emitter) {
+
+			String uadServiceKey = (String)serviceReference.getProperty(
+				"uad.service.key");
+
+			System.out.println("################## Heyo, registered");
+
+			if (Validator.isNotNull(uadServiceKey)) {
+				emitter.emit(uadServiceKey);
+			}
+		}
+
+	}
+
+}
