--- modules/apps/foundation/user-associated-data/user-associated-data-web/bnd.bnd
@@ -1,6 +1,0 @@
-Bundle-Name: Liferay User Associated Data Web
-Bundle-SymbolicName: com.liferay.user.associated.data.web
-Bundle-Version: 1.0.0
-Export-Package:\
-	com.liferay.user.associated.data.web.model,\
-	com.liferay.user.associated.data.web.service.registry
