--- modules/apps/foundation/user-associated-data/user-associated-data-web/build.gradle
@@ -1,11 +1,0 @@
-dependencies {
-	compileOnly group: "org.osgi", name: "org.osgi.core", version: "6.0.0"
-
-	provided group: "biz.aQute.bnd", name: "biz.aQute.bndlib", version: "3.1.0"
-	provided group: "com.liferay", name: "com.liferay.osgi.service.tracker.collections", version: "2.0.0"
-	provided group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "2.0.0"
-	provided group: "com.liferay.portal", name: "com.liferay.util.taglib", version: "2.0.0"
-	provided group: "javax.portlet", name: "portlet-api", version: "2.0"
-	provided group: "javax.servlet", name: "javax.servlet-api", version: "3.0.1"
-	provided group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
-}
