--- modules/apps/foundation/user-associated-data/user-associated-data-web/src/main/java/com/liferay/user/associated/data/web/model/impl/UADServiceImpl.java
@@ -1,67 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.user.associated.data.web.model.impl;
-
-import com.liferay.user.associated.data.web.model.UADAsset;
-import com.liferay.user.associated.data.web.model.UADService;
-
-import java.util.List;
-
-/**
- * @author William Newbury
- */
-public abstract class UADServiceImpl implements UADService {
-
-	@Override
-	public abstract void autoAnonymize(UADAsset uadAsset);
-
-	@Override
-	public void autoAnonymizeAll(long userId) {
-		for (UADAsset uadAsset : getUADAssets(userId)) {
-			autoAnonymize(uadAsset);
-		}
-	}
-
-	@Override
-	public long count(long userId) {
-		List<UADAsset> userIdUADAssets = getUADAssets(userId);
-
-		return userIdUADAssets.size();
-	}
-
-	@Override
-	public abstract void delete(UADAsset uadAsset);
-
-	@Override
-	public void deleteAll(long userId) {
-		for (UADAsset uadAsset : getUADAssets(userId)) {
-			delete(uadAsset);
-		}
-	}
-
-	@Override
-	public abstract void export(UADAsset uadAsset);
-
-	@Override
-	public void exportAll(long userId) {
-		for (UADAsset uadAsset : getUADAssets(userId)) {
-			export(uadAsset);
-		}
-	}
-
-	@Override
-	public abstract List<UADAsset> getUADAssets(long userId);
-
-}
